move JsonNames implementation to internal package
